requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bootstrap5==0.1.dev1
Werkzeug==3.1.3
SQLAlchemy==2.0.40
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
stripe==12.1.0
email_validator==2.2.0
dnspython==2.7.0
wtforms==3.2.1
psycopg2-binary
